+++ calibre-plugin/config.py
@@ -25,6 +25,8 @@ class PreferenceKeys:
     PREFER_OPEN_FORMATS = "prefer_open_formats"
     MAIN_UI_WIDTH = "main_ui_width"
     MAIN_UI_HEIGHT = "main_ui_height"
+    TAG_EBOOKS = "tag_ebooks"
+    TAG_MAGAZINES = "tag_magazines"
 
 
 class PreferenceTexts:
@@ -34,6 +36,10 @@ class PreferenceTexts:
     HIDE_EBOOKS = _("Hide Ebooks")
     HIDE_BOOKS_ALREADY_IN_LIB = _("Hide books already in library")
     PREFER_OPEN_FORMATS = _("Prefer Open Formats")
+    TAG_EBOOKS = _("Tag downloaded ebooks with")
+    TAG_EBOOKS_PLACEHOLDER = _("Example: library,books")
+    TAG_MAGAZINES = _("Tag downloaded magazines with")
+    TAG_MAGAZINES_PLACEHOLDER = _("Example: library,magazines")
 
 
 PREFS = JSONConfig(f"plugins/{PLUGIN_NAME}")
@@ -46,6 +52,8 @@ PREFS.defaults[PreferenceKeys.HIDE_BOOKS_ALREADY_IN_LIB] = False
 PREFS.defaults[PreferenceKeys.PREFER_OPEN_FORMATS] = True
 PREFS.defaults[PreferenceKeys.MAIN_UI_WIDTH] = 0
 PREFS.defaults[PreferenceKeys.MAIN_UI_HEIGHT] = 0
+PREFS.defaults[PreferenceKeys.TAG_EBOOKS] = ""
+PREFS.defaults[PreferenceKeys.TAG_MAGAZINES] = ""
 
 
 class ConfigWidget(QWidget):
@@ -54,7 +62,9 @@ class ConfigWidget(QWidget):
         self.layout = QGridLayout()
         self.setLayout(self.layout)
         label_column_widths = []
+        widget_row_pos = 0
 
+        # Setup Status
         self.libby_setup_status_lbl = QLabel(
             _("Libby is configured.")
             if PREFS[PreferenceKeys.LIBBY_TOKEN]
@@ -64,54 +74,106 @@ class ConfigWidget(QWidget):
             "font-weight: bold; "
             f'color: {"#00D228" if PREFS[PreferenceKeys.LIBBY_TOKEN] else "#FF0F00"};'
         )
-        self.layout.addWidget(self.libby_setup_status_lbl, 0, 0)
-        label_column_widths.append(self.layout.itemAtPosition(0, 0).sizeHint().width())
+        self.layout.addWidget(self.libby_setup_status_lbl, widget_row_pos, 0)
+        label_column_widths.append(
+            self.layout.itemAtPosition(widget_row_pos, 0).sizeHint().width()
+        )
+        widget_row_pos += 1
 
+        # Libby Setup Code
         self.libby_setup_code_lbl = QLabel(
             PreferenceTexts.LIBBY_SETUP_CODE
-            + ' [<a style="padding: 0 4px;" href="https://help.libbyapp.com/en-us/6070.htm"> ? </a>]:'
+            + ' [<a style="padding: 0 4px;" href="https://help.libbyapp.com/en-us/6070.htm"> ? </a>]'
         )
         self.libby_setup_code_lbl.setTextFormat(Qt.RichText)
         self.libby_setup_code_lbl.setOpenExternalLinks(True)
-        self.layout.addWidget(self.libby_setup_code_lbl, 1, 0)
-        label_column_widths.append(self.layout.itemAtPosition(0, 0).sizeHint().width())
-
+        self.layout.addWidget(self.libby_setup_code_lbl, widget_row_pos, 0)
+        label_column_widths.append(
+            self.layout.itemAtPosition(widget_row_pos, 0).sizeHint().width()
+        )
         self.libby_setup_code_txt = QLineEdit(self)
         self.libby_setup_code_txt.setPlaceholderText(
             PreferenceTexts.LIBBY_SETUP_CODE_DESC
         )
         self.libby_setup_code_txt.setInputMask("99999999")
         self.libby_setup_code_txt.setText(PREFS[PreferenceKeys.LIBBY_SETUP_CODE])
-        self.layout.addWidget(self.libby_setup_code_txt, 1, 1, 1, 1)
+        self.layout.addWidget(self.libby_setup_code_txt, widget_row_pos, 1, 1, 1)
         self.libby_setup_code_lbl.setBuddy(self.libby_setup_code_txt)
+        widget_row_pos += 1
 
-        self.hide_magazines_checkbox = QCheckBox(PreferenceTexts.HIDE_MAGAZINES, self)
-        self.hide_magazines_checkbox.setChecked(PREFS[PreferenceKeys.HIDE_MAGAZINES])
-        self.layout.addWidget(self.hide_magazines_checkbox, 2, 0)
-        label_column_widths.append(self.layout.itemAtPosition(2, 0).sizeHint().width())
+        # Tag Ebooks
+        self.tag_ebooks_lbl = QLabel(PreferenceTexts.TAG_EBOOKS)
+        self.layout.addWidget(self.tag_ebooks_lbl, widget_row_pos, 0)
+        label_column_widths.append(
+            self.layout.itemAtPosition(widget_row_pos, 0).sizeHint().width()
+        )
+        self.tag_ebooks_txt = QLineEdit(self)
+        self.tag_ebooks_txt.setPlaceholderText(PreferenceTexts.TAG_EBOOKS_PLACEHOLDER)
+        self.tag_ebooks_txt.setText(PREFS[PreferenceKeys.TAG_EBOOKS])
+        self.layout.addWidget(self.tag_ebooks_txt, widget_row_pos, 1, 1, 1)
+        self.tag_ebooks_lbl.setBuddy(self.tag_ebooks_txt)
+        widget_row_pos += 1
 
+        # Tag Magazines
+        self.tag_magazines_lbl = QLabel(PreferenceTexts.TAG_MAGAZINES)
+        self.layout.addWidget(self.tag_magazines_lbl, widget_row_pos, 0)
+        label_column_widths.append(
+            self.layout.itemAtPosition(widget_row_pos, 0).sizeHint().width()
+        )
+        self.tag_magazines_txt = QLineEdit(self)
+        self.tag_magazines_txt.setPlaceholderText(
+            PreferenceTexts.TAG_MAGAZINES_PLACEHOLDER
+        )
+        self.tag_magazines_txt.setText(PREFS[PreferenceKeys.TAG_MAGAZINES])
+        self.layout.addWidget(self.tag_magazines_txt, widget_row_pos, 1, 1, 1)
+        self.tag_magazines_lbl.setBuddy(self.tag_magazines_txt)
+        widget_row_pos += 1
+
+        # Hide Ebooks
         self.hide_ebooks_checkbox = QCheckBox(PreferenceTexts.HIDE_EBOOKS, self)
         self.hide_ebooks_checkbox.setChecked(PREFS[PreferenceKeys.HIDE_EBOOKS])
-        self.layout.addWidget(self.hide_ebooks_checkbox, 3, 0)
-        label_column_widths.append(self.layout.itemAtPosition(3, 0).sizeHint().width())
+        self.layout.addWidget(self.hide_ebooks_checkbox, widget_row_pos, 0)
+        label_column_widths.append(
+            self.layout.itemAtPosition(widget_row_pos, 0).sizeHint().width()
+        )
+        widget_row_pos += 1
 
+        # Hide Magazine
+        self.hide_magazines_checkbox = QCheckBox(PreferenceTexts.HIDE_MAGAZINES, self)
+        self.hide_magazines_checkbox.setChecked(PREFS[PreferenceKeys.HIDE_MAGAZINES])
+        self.layout.addWidget(self.hide_magazines_checkbox, widget_row_pos, 0)
+        label_column_widths.append(
+            self.layout.itemAtPosition(widget_row_pos, 0).sizeHint().width()
+        )
+        widget_row_pos += 1
+
+        # Prefer Open Formats
         self.prefer_open_formats_checkbox = QCheckBox(
             PreferenceTexts.PREFER_OPEN_FORMATS, self
         )
         self.prefer_open_formats_checkbox.setChecked(
             PREFS[PreferenceKeys.PREFER_OPEN_FORMATS]
         )
-        self.layout.addWidget(self.prefer_open_formats_checkbox, 4, 0)
-        label_column_widths.append(self.layout.itemAtPosition(4, 0).sizeHint().width())
+        self.layout.addWidget(self.prefer_open_formats_checkbox, widget_row_pos, 0)
+        label_column_widths.append(
+            self.layout.itemAtPosition(widget_row_pos, 0).sizeHint().width()
+        )
+        widget_row_pos += 1
 
+        # Hide books already in library
         self.hide_books_already_in_lib_checkbox = QCheckBox(
             PreferenceTexts.HIDE_BOOKS_ALREADY_IN_LIB, self
         )
         self.hide_books_already_in_lib_checkbox.setChecked(
             PREFS[PreferenceKeys.HIDE_BOOKS_ALREADY_IN_LIB]
         )
-        self.layout.addWidget(self.hide_books_already_in_lib_checkbox, 5, 0)
-        label_column_widths.append(self.layout.itemAtPosition(5, 0).sizeHint().width())
+        self.layout.addWidget(
+            self.hide_books_already_in_lib_checkbox, widget_row_pos, 0
+        )
+        label_column_widths.append(
+            self.layout.itemAtPosition(widget_row_pos, 0).sizeHint().width()
+        )
+        widget_row_pos += 1
 
         label_column_width = max(label_column_widths)
         self.layout.setColumnMinimumWidth(1, label_column_width)
@@ -125,6 +187,9 @@ class ConfigWidget(QWidget):
         PREFS[
             PreferenceKeys.HIDE_BOOKS_ALREADY_IN_LIB
         ] = self.hide_books_already_in_lib_checkbox.isChecked()
+        PREFS[PreferenceKeys.TAG_EBOOKS] = self.tag_ebooks_txt.text().strip()
+        PREFS[PreferenceKeys.TAG_MAGAZINES] = self.tag_magazines_txt.text().strip()
+
         setup_code = self.libby_setup_code_txt.text().strip()
         if setup_code != PREFS[PreferenceKeys.LIBBY_SETUP_CODE]:
             # if libby sync code has changed, do sync and save token